Add (temporary) resume to video playback (#7)
This is a quality-of-life improvement. When exiting a playing video,
pass the current position back to the `VideoDetailsFragment`, so the
user can optionally resume the playback.

This helps when seeking in a video if you accidentally press `back` too
many times (trying to get rid of the controls), you can resume the video
playback. This even seems to work despite #6.

The position is *not* persisted and is lost when the
`VideoDetailsFragment` is removed (e.g. pressing back on the details
page).

Some apps remember playback positions persistently, which might be a
good future enhancement?