Override video details max lines (#38)
Increase the max lines from the default of 5 to 18. 18 seems to be the
limit that fits on the screen well.

Ideally, the details would be "clickable" which would expand to the full
description, but I haven't figured out how to add that yet. It would
likely require implementing a custom
`VideoDetailsFragment`/`DetailsSupportFragment`.