Add rating to scene details and allow updating the rating (#118)
Adds the scene's rating to the scene details page. If the type is
`start`, a `RatingBar` (five stars) is used, otherwise for `decimal` a
`SeekBar` is used.

You can highlight the rating and move it left/right and then press
enter/OK to set the new rating.

Like usual, the UI is a bit rough, but it's a great feature to add I
think and work on refining later.