Add option to scroll long scene details text (#123)
Adds a preference (enabled by default) which adds scrolling to scenes
with long details. Long details are about ~18 lines or longer.

When enabled, the entire details section is scrollable and must be
scrolled to the bottom to move to the rows below. If no scrolling is
needed, then it acts as if disabled.

When disabled, the details section is truncated to whatever fits in the
view.