Load main page content in order asychronously (#140)
Replaces #139 

After https://github.com/damontecres/StashAppAndroidTV/pull/131 &
https://github.com/damontecres/StashAppAndroidTV/pull/133, the rows on
the main page load asynchronously. And the first one to load grabs the
focus of the page which is not ideal.

This PR refactors so that all of the data is fetch asynchronously, but
added to the UI in order.

Additionally, in the event of failures, focus will be moved to the top
bar. Before focus would be lost and you have to press up to get to the
bar to scroll over the settings.