Fixes to scene details UI glitches (#135)
In #81, scene details were updated to fetch during `onResume` which
meant that all of the markers, performers, etc would be reset. This
meant everything could flash for a second. Also meant that the selected
item would revert to the beginning of the list.

So this PR moves the initial fetch of scene data to `onViewCreated` and
does a refresh in `onResume` so those UI glitches won't happen.

This PR also moves some of the UI initialization to _before_ fetching
the scene data and removes the temporary background image so there isn't
a strange color swap.