Fix blue flash when card video starts on some devices (#167)
Image's don't have a background so that transparent images show the card
background.

Since the entire card's background changes when selected, this means the
selected color will flash for a moment when switching to the video.

Changing the background to black before switching will remove that flash
since the video background is also black. And then changing it back
after will use the transparency.