Convert IDs to Strings and other code cleanup (#191)
https://github.com/stashapp/stash/pull/4563 (added in `v0.25.0`) updated
the remaining find queries to use `ID` instead of `Int`. The `ID` type
in graphql is serialized as a String, but the server implements them as
auto increment integers in the database.

This PR updates all of the ID references to Strings for any potential
future proofing.

Note: This PR does not updated the find queries used by this app to use
the ID fields. See deferred list in #132.