Fix crash at end of video playback (#193)
Fixes #190 

Refactor the `saveSceneActivity` timer to use an actual Timer instead of
a coroutine delay. The coroutine delay doesn't guarantee that it doesn't
block the thread it's running on (which was the main/UI thread in this
case).

The actual Timer uses a separate thread which won't ever block the main
thread.