Fix crashes due to new timer thread (#198)
Follow up to #193

Since the new thread added in #193 is no longer tied to the fragment's
lifecycle, it continues to run even after the playback activity is
finished. This results in a crash when the timer thread tries to access
the fragment's view lifecycle to make the network call.

This PR ensures that the timer is cancelled when the playback activity
ends.