Overhaul scene text overlay (#201)
Closes #199 
Closes #89

Changes the text overlays for scenes to use views instead of image
edits. This means the overlay will be hidden when selecting a card.

Also, ratings & studio image/text are added. Plus, the rating overlay is
added to all of the data types that support it.

Note: the rating can be hidden in settings since the background color is
a little out-of-place. It is 50% transparent though which softens the
background color.

<img width="623" alt="textoverlay"
src="https://github.com/damontecres/StashAppAndroidTV/assets/154766448/a96ac86a-775a-4bbf-b302-4b109c2da6fc">