Better playback D-Pad non-skipping behavior (#211)
Closes #179 

This PR changes the left/right D-Pad behavior during playback. If D-Pad
skipping is _disabled_ in settings, then when left/right is pressed
during playback and the controls are _not_ showing, the controls will be
shown, focus given to the seek bar, and the video preview (if available)
will be shown.

Previously, any D-Pad button would just show the controls and focus on
the Play/Pause button. Now, this is the behavior with D-Pad
up/down/center/etc only.

If D-Pad skipping is _enabled_, then D-Pad left/right will skip
back/forward as before.