Better popup width caclulation (#215)
A popup's width is tied to the anchor view which can be pretty small. So
this PR adds a function to measure the width of the text for each item
in the popup so the popup can be sized better.

The function can also calculate the max width based on the window size.