Simplify settings (#228)
Closes #226 

Moves about half of the settings to an "Advanced" screen. This makes the
most important settings easier to navigate.

Implementing #225 will make this even better because 6 entries in
settings are devoted to server settings.

This PR also changes all of the `findPreference` calls to use the `!!`
not null assertion which ensures any configuration on either setting
screen is applied and saved.