Add default long click listener to cards (#231)
Closes #151

If a presenter doesn't have a `LongClickCallBack`, a default one is used
with a single `Go To` option which, when selected, is the same as
clicking the card.

Before this, long clicking a card without a `LongClickCallBack` would do
the same as a normal click. So this PR adds an extra step there, but I
feel like that's more intuitive. It felt weird to long click something
for it to just act like a regular click, except for some cards it
doesn't.