Add network timeout setting for large servers (#253)
Adds a setting to the advanced page to change the network timeout. The
previous default was 10s (now 15s). Setting to zero will disable the
timeout all together. Large servers can take longer than 10 seconds to
run queries.

The setting is also clickable to test the server connection so you don't
have to go back and forth to find a good value.

I tested this with the emulator's latency/speed controls.