Add a new UI for server configuration (#262)
Closes #225 

Adds a guide step UI for 1) the initial setup on a new install and 2)
adding, removing, or switching servers.

The setup UI guides a new user in setting their server URL, API key (if
needed), and adding a PIN (if they want). This is much better UX than
getting dropped into the main page with no content and an error message.

Similarly, 5 preferences items are condenses to 2 for adding, removing,
or switching servers. The UX similarly guides the user when adding a new
server.

I did a little testing for pasting text from an phone, but I suspect
there may be similar issues as #63.

Also, the test query is changed so it runs fewer SQL queries on the
server. This may help larger server to connect.