Improvements to initial setup (#267)
Follow up to #225 & #262 

The initial setup is now broken into individual steps:
1. Server URL
2. If needed: trust self-signed certs
3. If needed: Server API key
4. Set a PIN, optionally

Steps 2 & 3 will only be presented if needed. So if the server URL uses
`http` or `https` with a trusted certificate, then step 2 will be
skipped. Similarly, if the server doesn't require auth, step 3 will be
skipped.

In step 4, the user can set & confirm a PIN, or skip it.

These individual steps will better guide a new user on what settings
they need to configure.

Finally, the messages along the way are improved with better
information.