Transcoding updates to #273 (#274)
A follow up to #273

Adds the container format to the transcode decision. The defaults are
what ExoPlayer should support out of the box, but users can override
this in settings.

Also more video & audio codec override options, but most are pretty
uncommon.

I tested this PR, along with #273, on two NVIDIA Shields (2017 & 2019
non-Pros), Fire TV Stick Lite, and the emulator. Seems to work pretty
well overall.