Speed improvements for card view layout & main page load time (#289)
`PlayerView`s for a card are now inflated on demand. Since this is a
somewhat expensive view to create, this speeds up the layout a bit
(micro benchmarks show up to 50%, but more like ~20% on average). The
trade off is it takes an extra 50-100ms to start a video preview now.

The icon row is switched to a `ConstraintLayout` with fewer nested
views. This saves a small amount of time.

The main page now starts loading data during `onCreate` instead of
`onResume` which saves a small amount of time (100-300ms maybe) while
the page's view is being inflated.