A possible fix for #288 (#293)
Related to #288 

I haven't been able to reproduce the issue in #288, so I don't know if
this PR will fix it, but this PR doesn't break anything either.

Specify the focus-down views for the performer tabs to be the
`ViewPager` which holds the scene grid. This _should_ mean that when a
tab is focused, pressing D-Pad down will _always_ focus down to the
`ViewPager`.