Better support for direct playing `m3u8`/`HLS` files (#305)
Closes #302 

Adds `hls` as a container format. Additionally, if direct streaming an
`hls` (or `dash`) file, explicitly set the mime type so that ExoPlayer
plays it properly.

Also, if the server requires an API key, the API key header will only be
set for requests to the server and not other domains or IP addresses.