Add a page for galleries (#314)
Related to #312 

Replaces the grid of images in a gallery with a page similar to the
performer page. The left side shows details about the gallery and the
right has tabs for images, scenes, performers, & tags.

This allows for viewing the gallery details and other related items.

The performer and tag fetching is not very efficient. The gallery data
is queried three times, but all of the data is available in each query.
Will need to follow up on this.

Also, I found some SVGs on my test server have unusual bounds and are
rendered much larger than expected. This also needs follow up.