Experimental: Play a list of markers for the specified duration as a compilation (#327)
This is an experimental feature!

On a grid of scene markers, if experiment features are enabled, there is
now a "Play All" button. Clicking it gives a duration option.

Then all of the scene markers will play back-to-back. Each scene will
only play for the specified duration. Errors in playback are shown and
the scene skipped.

A few problems:
- ~Network access for playing videos is changed which breaks
https://github.com/damontecres/StashAppAndroidTV/pull/305/commits/981ff0638090f289e827d81d5c4eefedee515e59~
- ~Without this change, an out-of-memory error will occur after a few
markers play, so something might be wrong with caching?~
- This does use more memory, but the safety is required. The memory
issues only occur on the very short 3 second clips
- Suffers from #326 as well
- ~Lots of duplicated code between the playback fragments~

Refactors `PlaybackExoFragment` into an abstract parent class
`PlaybackFragment` to reduce code duplication.