Add support for image clips (#333)
Closes #331

Adds support for video files added as image clips. You can scroll
through images/clips as before with left/right (unless the overlay is
showing). The overlay for image clips is basically the scene playback
controls.

There's no ability to rate an image clip since they use a different
overlay. This will have to be a follow up I think.

This was pretty straightforward thanks to the refactoring in #327. Also,
this is affected by #328.