Send logs & crash reports to stash server (#339)
Because gathering logs and information about app crashes is so
difficult, this PR adds the ability to gather that information and send
it to the user's Stash server if the companion plugin is installed.

The logs are only ever sent to the user's own stash server to help
protect privacy.

If the app crashes, a dialog will appear asking whether or not to send a
crash report to the server. The report will be logged server-side in Log
tab of Settings (eg http://192.168.1.5:9999/settings?tab=logs). This can
be disabled in advanced settings (ie the dialog will never appear).

PR to add the plugin to the community scripts:
https://github.com/stashapp/CommunityScripts/pull/381