Fix file path parsing for galleries (#341)
Fixes #338 

If a gallery doesn't have a title, it should fall back to the file or
folder name. The parsing for the file/folder name used `Uri.parse(..)`
which is definitely not correct and resulted in a crash for some paths.

I tested some variations with `File(..).name`, but decided adapt the
[regex](https://github.com/stashapp/stash/blob/v0.26.2/ui/v2.5/src/utils/text.ts#L218-L221)
that the server uses instead for this purpose.