Add ability to rotate scene's video playback (#347)
Closes #202

Adds ability to rotate the video during scene playback. The rotation is
persisted by default so that subsequent plays can use the same rotation.

The rotation value is persisted locally on the Android TV device and not
shared with other devices or clients. This persistence can be disabled
in advanced settings.

Dev: also bump some dependencies versions