Add ability to shift a marker's timestamp (#361)
Adds an action to the marker details page to select the number of
seconds to shift the marker's timestamp by.

This is especially useful if creating markers in the app since it's
difficult to exactly get the precise timestamp. This feature allows for
refining the value.

The range is equal to the smaller of the skip back or skip forward
settings.