Clean up apks after update (#378)
After the app is updated, clean up the previously downloaded APKs.

The content resolver will stop downloading after the 32nd update and you
would have to manually delete the APKs to update again. Also while the
APKs aren't huge, there's usually not a ton of storage on Android TV
devices, so it is a good idea to clean things up.

This should work retroactively to clean up all previous downloads, not
just the latest.