Fix crashes from very large scrolls (#393)
Scrolling a far distance quickly could cause a crash. This is because it
would try to load placeholders, but the presenters don't handle null
items (placeholders are null).

So this PR updates the presenters to handle nulls, but also tweaks large
scrolls to jump instead of smooth scroll.

Also the playlist list will now jump to the currently playing item
instead of scrolling.