Large refactor to reduce usage of `Context` (#417)
Follow up to #415 to further reduce usage of `Context` everywhere.

Another change is removing showing toasts in `QueryEngine` and
`MutationEngine`. It wasn't used often and also it pretty much violated
the single responsibility principle. The toasts are better handled in
response to thrown exceptions.

Many of the changes make it a lot easier to mock and write unit tests in
the future.