Small UI/UX tweaks to setup, filter creation, & non-TV devices (#423)
A few UI/UX tweaks to the initial setup screens

Also changes the text about create filters

Finally, while the app does not explicitly support non-TV Android
devices, there's no need to overly enforce that. So this PR relaxes a
few manifest and touchscreen settings which makes the experience a bit
nicer, but still not supported!