Add ability to filter from a tab page (#437)
Adds a button to tabs to create a filter from it. This just saves some
navigation and clicks if you want to filter/search for something
specific.

For example, if viewing the galleries tab of a performer, clicking the
`Filter` button will allow for creating a filter for galleries including
that performer.