Use server's default filter for group scenes (#452)
If the server has set a server-side default filter for group scenes, the
app will now use it.

Most of the changes are refactoring to make sure server preferences can
be available before tabs are loaded. The view model now has
`currentServer` and `tabs`. The tabs are added based on the `tabs` view
model field. So if a page needs server preferences, it can observe the
`currentServer` before setting `tabs`.

Will need to follow up on various other default filters too.