Use server provided page sorts for tabs if available (#457)
Follow up to #452
Part of #453

If the user has saved a default sort on a data type page (eg a
performer's page scene list or a studio's page image list), the app will
now use that. The app will fallback to the standard default sort if
needed.

For now only sorting is used from the server. A follow up will include
other (object) filters.

The list is derived from
https://github.com/stashapp/stash/blob/v0.27.2/ui/v2.5/src/components/List/views.ts