Updates/fix to setup process (#469)
Related to #468 

Hopefully resolves issues with entering the API key.

Adds submit buttons so that if the UI gets in an unusual state, there's
still a way to submit and move to the next step.

Reverts the multi-line change from #458 since that seems to cause lots
of issues. I think another issue was `hasNext(true)` was removed in #458
so there wasn't the same hints sent to keyboard.

Also adds a hack that might help touch screen devices to enter the info
by examining the underlying edit text views.