Switch grid paging to custom implementation (#472)
After wasting a lot of time trying to get large jumping working with
Paging 3 and reading about how most people avoid it, this PR replacing
it with a custom `ObjectAdapter`. This means the PR needs a LOT of
testing.

The new `PagingObjectAdapter` caches a few pages and prefetches as the
user moves up/down. It also supports fetching/jumping to an arbitrary
position.

Some downsides:
1. The behavior might be affected if the number of items changes on the
server. This needs further testing.
2. It needs to get the total number of items before loading any data
which adds a small delay, but it might be significant for large servers