Jump to a letter on a grid page (#473)
Closes #317 

When sorting by `name` for Groups, Performers, Studios, or Tags, a list
of letters will show and can be clicked to jump to the first item
starting with that letter.

This process will probably be slow-ish for very large servers, but my
tests on the emulator with 7000+ items, the query was <30ms and the
whole jump took ~1.5s.

This PR also adds loading indicator for all grid pages to make it clear
what's happening.

This PR also fixes counting the tags on a studio.

Only supports the English alphabet for now.