Add page up/down buttons to grids (#475)
Adds two page up and two page down buttons. They jump a little and jump
a lot.

The amount of items jumped scales up based on the number of columns and
total number of items. I'm not sure if having the number vary is a good
idea or not.

The alphabet buttons are moved to the left side as well.