Adjust card size and move letter jump buttons to the right (#477)
Follow up to #473, #474, & #475

The paging buttons are on the left size. The alphabet/letter buttons
appear on the right if applicable.

This change required modifying the cards slightly so everything fits
well without a horizontal scroll. As a follow up, I want to find a way
for the cards to automatically resize themselves instead of the
hard-coded sizes.


![performers_jump_buttons](https://github.com/user-attachments/assets/54530936-d554-4cce-a6bc-bd539a9b8a8a)