Cleaner loading of initial grid page items (#485)
After switching to the new pager (#472), when loading the first page of
items on a grid, it would flash and resize awkwardly because the first
several items would report null while the actual data is being fetched.

So this PR moves things around so that the grid initially thinks it has
zero items. Then the count and first page of data is fetched before
notifying the UI that there are items now.