Rewrite the app to use a single activity (#501)
Refactors and rewrites significant portions of the app so that
everything runs within a single activity. This is more in line with
modern Android development. The MVVM model is used a lot more as well.

Also adds a navigation component which makes it much easier to figure
out the required arguments for a page. This also removes a lot of
serialization partial objects and relies on item IDs instead.

There are minor user facing changes such as:
- Updating the app and clicking "open" at the end actually re-opens the
app
- Some better animations between pages or when scrolling
- Minor changes to a couple pages

There are a huge amount of changes and this needs lots of testing.