Save applied image effects (#508)
Similar to the video effect filters, image effect filters can now be
saved and will be applied automatically when viewing the image.

One downside is that the effects are applied _after_ the image is
loaded, so there's a flash of the original image before the effects are
applied. I experimented with applying the effects on load, but then
further effect changes effectively double up, so need more investigation
there.

Also, the rotation & flip is not saved yet.

Dev: adds a DB migration from 4 to 5