Fixes when items can't be removed after going back to scene page (#509)
`RemoveLongClickListener` takes `CoroutineScope` and the details pages
use `viewLifecycleOwner.lifecycleScope` for this.

But once navigating away from the details page, that scope is ended
because the view is gone. This PR refactors to use a function to
retrieve the scope when needed, so the details fragment can provide the
current `viewLifecycleOwner.lifecycleScope`.