Add search to grid pages (#510)
Adds a search button to the grid pages which expands to enter a query.

To make the UI more consistent and workable for the search bar, the
buttons on the filter page are moved from the upper right corner to
below the stash icon on the left. This is the same as the tabbed grid
pages.