Long press back to go to previous position (#524)
When navigating a grid page, you can long press the back button to go
back to the previous spot.

For example, short pressing back jumps to the top, you can long press
back to go back where you were.

Dev: adds a pre-commit check to check if `DEBUG` is true anywhere since
it should only be when debugging a class.