Prevent content showing briefly before PIN screen (#530)
Previously, when returning to the app, the previous page/content might
flash briefly before the PIN screen is shown. This PR switches to the
PIN when the app is pausing to prevent this.

Also make sure that if a PIN is added, it is used immediately instead of
needing the app to restart before taking effect.