Fix cases where ExoPlayer is not released (#538)
Part of #501 changed how some instances of ExoPlayer was created, but
did not always clean up correctly. I think this lead to a memory leak
and eventually the app would crash or lock up.

Before #501, the `StashApplication` released the player whenever an
activity paused/stopped, so the player was always eventually cleaned up.
This PR ensure the player will be cleaned up when the app goes into the
background or when a new player is required due to different settings.

I think will solve #537 as well.