Always reset player for playback (fix no audio) (#540)
Follow up to #538 

Always recreate the player when starting playback for scenes or markers.
This ensures a fresh newly configured player without any of the
overrides used for the cards' previews.

Before this PR, if card previews was enabled and card preview audio was
disabled, then scene playback would start with audio disabled.

Also refactors to force correct initialization of the player in
`PlaybackFragment`.