Fix more memory leaks (#548)
Continuing from #501 and #538, this PR fixes more memory leaks.

The leaks in Scene/Image details were due to various lambdas in
action/click listeners.

Additionally, the `onlyScrollIfNeeded` functions added a global layout
listener that references the `ScrollView`, which in turn references the
fragment, but the listener was never removed. That listener is
refactored out now.