Allow creating items when search has non-conflicting results (#565)
Previously when searching to add a group, performer, or tag, you could
create a new one that matched the query only if there were no results.
This meant it wouldn't be possible to create one that closely matched an
existing item.

For example, if "Quick Brown Fox" was an existing tag, a search for
"Quick" would match and would prevent creating a new one called "Quick".

This PR allows for creating the items unless the query directly matches
an existing name or alias.