Activity Tracking improvements (#571)
The activity tracking will now use a singular `Timer` instead of
creating a new one for every playback. This will prevent lots of useless
threads from being created.

The tracking now uses millisecond precision when saving to the server.

Activity tracking is disabled when playing a playlist of markers and the
duration for each marker is less than 2 minutes. This saves resources
(which are constrained with quick media changes) but I also think it
doesn't really make sense to track activity when viewing markers in
quick secession.