Show scene screenshot when updating a marker's timestamp (#566)
Instead of only being able to shift a marker's timestamp by +/- 10
seconds, now the timestamp can be set to any value.

There's now a picker for hours, minutes, seconds, & milliseconds.
Additionally, when changing the timestamp, the app will attempt to
extract a screenshot for the specified time.


![shift_marker](https://github.com/user-attachments/assets/e8f7c5a3-d424-488d-b4fa-bc5aa2eb7576)

Note: This PR requires features in Stash v0.28.0 so it is targeting
`develop/v0.6.0`