Load thumbnail while waiting for full image (#579)
This PR causing the image view page to simultaneously load the image and
its thumbnail. Since thumbnails are much smaller for larger images, this
means the thumbnail will be displayed while the full image is being
downloaded.

This does cause a little hiccup in the UI when transitioning from
thumbnail to image due to either 1) the crossfade or 2) perceivable
differences between thumbnail & image (HDR or other image factors
maybe?)