Advanced settings to adjust image loading (#583)
Most users do not need to change anything here!

Adds an advanced preference to control the image disk cache size
allowing for extreme large images.

Also adds an advanced preference to control the number of threads used
to download images. This is mostly for testing.

Finally, adds a minor enhancement to cancel in-progress image loads for
card if the card is unbinding from the screen (ie don't download images
that won't be displayed).