Play all of the markers for a scene (#618)
This basically lets you play the "highlights" of a scene if you've set
up the markers for it!

Adds a card to the end of marker list for a scene if there are 2+
markers.

This will play all of the markers for the scene in order. If at least
one marker doesn't have an end time, it will prompt for duration to play
markers without end time. Otherwise, it immediately will play each
marker for its duration.

Having the play all at the end of the marker list can be annoying if
there's a lot of markers though.

Related to #604