Move play all markers to more menu (#619)
Updates #618 

Moves the `Play all markers` button to the `More` menu dialog. And since
that dialog had a lot of options now, all of the edit functions are
moved to a new `Edit` button (which isn't shown if read-only mode is
enabled).

Also, this PR fixes a bug where playback tracking would be enabled for
marker playlist which messes up the resume time for a scene.