Allow more than 100 items in a playlist (#621)
Playing a playlist will now support as many items as the device memory
can handle.

Like the old UI, as you advance in the playlist, new items will be
fetched and added to the playlist.

There's no limit, but eventually memory will run out because every item
in the playlist needs to be kept in memory.