Add option to show embedded subtitles when direct playing scenes (#624)
This PR adds the ability to select subtitle/caption tracks embedded in a
scene's file if the file is direct played. They are shown in addition to
any "sidecar" captions provided by the server.

Additionally, image subtitles (eg vobsub, pgs) are now supported. Though
the UI isn't always great.

As a consequence of this, a bug is fixed where captions may not show
when activated when there are embedded subtitles and sidecar captions.