Fix fast scrolling through a grid in new UI jumping to nav drawer (#647)
Previously if scrolling very fast down the grid, focus might jump over
to the nav drawer. This PR fixes this behavior.

This happened because items are lazy loaded from the server and a
placeholder card is shown while waiting for the result. This means the
user may be focused on a placeholder card when the data is loaded. In
that case, the card is replaced with the real card, but focus is not
maintained, so the app reverts focus to a default place, ie the nav
drawer.

This PR updates the cards to support a null object while waiting a
result from the server to replace the item not the card. Thus the same
card is used as both a placeholder and real and focus is maintained when
the data is loaded.