Add ability to create new studios (#649)
When setting the studio for a scene, adds ability to quickly create a
studio named as the search query.

Also, in the new UI, the option to create a studio was shown, but would
crash the app. So additionally this PR makes it so, when searching, data
types that don't support creation are not shown with the option to
create (markers, scenes, performers, & galleries).

This works in the new UI and legacy UI.