Automatically send error logs to the server (#655)
Closes #652

Updates the new UI to automatically log errors in the app on the server
(if the companion plugin is installed). This will make it easier to
report problems in the app.

It can be turned off in advanced settings in the app.

Also enhances the regular "Send Logs" option in settings to include
information about the device such as model and app version.