Better playback scene details navigating (#658)
#657 Added navigation from the scene details overlay during playback,
but it was implemented by go back from playback then navigating to the
specified item. This is not a good user experience at all because going
back from the selected item doesn't actually return the user to where
they were previously.

This PR instead remembers a few key playback values such as the current
playlist index and current scene position when navigating away. So now
when going back, playback can be resumed from it was left off. This is
much more intuitive and a much better UX.