Fixes random default sort in the legacy UI (#664)
Fixes #663 

In the legacy UI, ensure each time a data type with a default random
sort is opened, the sort is, in fact, randomized.

This was fixed in the new UI in #635 for #634

This PR also refactors some of the drawer page logic in the new UI so
that server-side changes in default filters are reflected on the next
configuration query.