Display marker info during playlist playback (#673)
Closes #670

When playing markers in a playlist, instead of showing the scene's info
on the player controls overlay, show details about the marker such as
its title, primary tag, duration, etc.

Also updates the playlist list items with the same info.

I don't think it makes sense to do the same for clicking a single marker
playback (eg not Play All), since that plays the full scene.