Long press Up/Down buttons to zoom in/out on images (#675)
Closes #668 

When viewing an image, long pressing D-Pad Up or Down will slowly zoom
the image in or out respectively.

I also experimented with allows for long pressing left & right to pan,
but while trying it, it made me also want to long press up/down to pan,
but that action zooms, not pans, so it felt awkward.

Note: repeated single presses of any D-Pad direction always pans the
image while zoomed in.