Adjust image zoom & pan (#679)
Follow up to #675

Adjusts the "center" of the zoom when zooming in and out. So zooming out
will also pan the image towards the image's center, so when zoomed to
full size, the image will be centered. Zooming in does not pan, so it
zooms towards the screen's center.

Also limits the max pan amount so you can't move the image off screen.