Experimental support for touchscreen devices (#696)
This is a work in progress!

This PR makes the minimal changes needed to the new UI so that
touchscreen devices, ie phones and tablets, work okay-ish.

The UI is not changed at all! It is **not** optimized for smaller
screens! It does **not** look like a phone app and does **not** follow
the guidelines for phones/tablets. But it works okay!

Note: this only works for the new UI! Touch gestures like swiping or
dragging do not work in most cases.