Maybe fix performer images not loading (#705)
~Downgrades Coil because seemingly starting with 3.1.0 (maybe caused by
https://github.com/coil-kt/coil/pull/2795?), sometimes on the performer
page the image won't show. I need to experiment more and make a
reproducer before filing a bug report though.~

It seems that disabling the cross fade solves the issue. But the cross
fade is used for studio images which don't seem to have the problem, so
I'm not sure if this is a true fix.

Also updates several dependencies