Add button to show scene description dialog (#718)
Adds an additional button under more to show the scene
details/description dialog. This is useful when the scene has no
description (so its not clickable) but you want to see the other info
that dialog has such as file info.

Also adds the full scene title to the dialog since it could be
truncated.

Also if there are no markers in the scene, remove the option to play
markers.